+++ src/pages/Home.tsx
@@ -1,38 +1,11 @@
 import { styled } from "styled-components";
 import MasonryLayout from "../components/layouts/MasonryLayout";
 import HeaderBar from "../components/layouts/HeaderBar";
-import axios from "axios";
-import { useEffect, useState } from "react";
-
-interface ITitle {
-  title?: string;
-}
+import articleHooks from "../hooks/articleHooks";
 
 function Home() {
-  const [articleData, setArticleData] = useState();
-
-  const getAllArticle = async () => {
-    try {
-      const res = await axios.get("https://team7.collab-pinterest.p-e.kr/pin");
-
-      setArticleData(res.data.data.pins);
-    } catch (err) {
-      console.log(err);
-    }
-  };
-  const searchArticle = async (title: ITitle) => {
-    try {
-      const res = await axios.get(`https://team7.collab-pinterest.p-e.kr/pin/?title=${title}`);
-      setArticleData(res.data.data.pins);
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  useEffect(() => {
-    getAllArticle();
-    return () => {};
-  }, []);
+  
+  const { articleData, getAllArticle, searchArticle } = articleHooks();
 
   return (
     <>
